HW04 Replication: запуск mysql репликации в режиме master-slave

+++ src/main/java/ru/apolyakov/social_network/controller/UserController.java
@@ -59,7 +59,7 @@ public class UserController {
         if (userId == null) {
             return "redirect:/users";
         }
-        Long currentUserId = userService.getCurrentUserId();
+        Integer currentUserId = userService.getCurrentUserId();
         if (currentUserId == null || currentUserId.intValue() == userId) {
             return "redirect:/index";
         }
